Bug fixing ->Search Ui Update -> Search  by location or code -> update the button font -> fix other Ui Issue

+++ app/src/main/java/app/denhan/model/jobs/MaintenanceJob.kt
@@ -11,7 +11,7 @@ data class MaintenanceJob(
     val f_start_time: String,
     val id: Int,
     val job_detail: String,
-    val labour_charges: Int,
+    val labour_charges: Double,
     val maintenance_id: Int,
     val modified_date: String,
     val start_time: String,
